Stop extending canvas when converting SVG to PNG
A larger canvas won't prevent clipping; if the inner elements of an SVG are
rendered larger than the specified width and height, they are still clipped at
the designated boundary. So we were extending and trimming for naught, and
wasting cycles doing so.

The optipng call is also parallelized now.